# Optimize __ziplistCascadeUpdate algorithm