--- a/src/ziplist.c
+++ b/src/ziplist.c
@@ -1439,6 +1439,17 @@ int ziplistTest(int argc, char **argv) {
     ziplistRepr(zl);
 
     zfree(zl);
+    {
+        char data[ZIP_BIG_PREVLEN];
+        zl = ziplistNew();
+        time_t start = time(NULL);
+        for (int i = 0; i < 100000; i++) {
+            zl = ziplistPush(zl, (unsigned char*)data, ZIP_BIG_PREVLEN-4, ZIPLIST_TAIL);
+        }
+        zl = ziplistPush(zl, (unsigned char*)data, ZIP_BIG_PREVLEN-3, ZIPLIST_HEAD);
+        printf("benchmark __ziplistCascadeUpdat elapsed %zd\n", time(NULL)-start);
+        zfree(zl);
+    }
 
     printf("Get element at index 3:\n");
     {