--- a/src/ziplist.c
+++ b/src/ziplist.c
@@ -2015,7 +2015,7 @@ int ziplistTest(int argc, char **argv) {
         verify(zl, e);
 
         assert(e[0].prevrawlensize == 1 && e[0].prevrawlen == 0);
-        compareHelper(zl, 'a', s1, 0);
+        assert(compareHelper(zl, 'a', s1, 0));
         ziplistRepr(zl);
 
         /* No expand. */
@@ -2023,10 +2023,10 @@ int ziplistTest(int argc, char **argv) {
         verify(zl, e);
 
         assert(e[0].prevrawlensize == 1 && e[0].prevrawlen == 0);
-        compareHelper(zl, 'b', s1, 0);
+        assert(compareHelper(zl, 'b', s1, 0));
 
         assert(e[1].prevrawlensize == 1 && e[1].prevrawlen == strEntryBytesSmall(s1));
-        compareHelper(zl, 'a', s1, 1);
+        assert(compareHelper(zl, 'a', s1, 1));
 
         ziplistRepr(zl);
 
@@ -2035,13 +2035,13 @@ int ziplistTest(int argc, char **argv) {
         verify(zl, e);
 
         assert(e[0].prevrawlensize == 1 && e[0].prevrawlen == 0);
-        compareHelper(zl, 'c', s2, 0);
+        assert(compareHelper(zl, 'c', s2, 0));
 
         assert(e[1].prevrawlensize == 5 && e[1].prevrawlen == strEntryBytesSmall(s2));
-        compareHelper(zl, 'b', s1, 1);
+        assert(compareHelper(zl, 'b', s1, 1));
 
         assert(e[2].prevrawlensize == 5 && e[2].prevrawlen == strEntryBytesLarge(s1));
-        compareHelper(zl, 'a', s1, 2);
+        assert(compareHelper(zl, 'a', s1, 2));
 
         ziplistRepr(zl);
 
@@ -2050,16 +2050,16 @@ int ziplistTest(int argc, char **argv) {
         verify(zl, e);
 
         assert(e[0].prevrawlensize == 1 && e[0].prevrawlen == 0);
-        compareHelper(zl, 'd', s2, 0);
+        assert(compareHelper(zl, 'd', s2, 0));
 
         assert(e[1].prevrawlensize == 5 && e[1].prevrawlen == strEntryBytesSmall(s2));
-        compareHelper(zl, 'c', s2, 1);
+        assert(compareHelper(zl, 'c', s2, 1));
 
         assert(e[2].prevrawlensize == 5 && e[2].prevrawlen == strEntryBytesLarge(s2));
-        compareHelper(zl, 'b', s1, 2);
+        assert(compareHelper(zl, 'b', s1, 2));
 
         assert(e[3].prevrawlensize == 5 && e[3].prevrawlen == strEntryBytesLarge(s1));
-        compareHelper(zl, 'a', s1, 3);
+        assert(compareHelper(zl, 'a', s1, 3));
 
         ziplistRepr(zl);
 
@@ -2069,13 +2069,13 @@ int ziplistTest(int argc, char **argv) {
         verify(zl, e);
 
         assert(e[0].prevrawlensize == 1 && e[0].prevrawlen == 0);
-        compareHelper(zl, 'd', s2, 0);
+        assert(compareHelper(zl, 'd', s2, 0));
 
         assert(e[1].prevrawlensize == 5 && e[1].prevrawlen == strEntryBytesSmall(s2));
-        compareHelper(zl, 'c', s2, 1);
+        assert(compareHelper(zl, 'c', s2, 1));
 
         assert(e[2].prevrawlensize == 5 && e[2].prevrawlen == strEntryBytesLarge(s2));
-        compareHelper(zl, 'a', s1, 2);
+        assert(compareHelper(zl, 'a', s1, 2));
 
         ziplistRepr(zl);