@antirez 2020-02-20
@neal-zhu thank you, this is interesting. However while the worst case of the algorithm is O(N^2), what is the normal time it is spending there? In other words, are you able to measure a real improvement here if you test Redis itself in a common workload? Thanks.