@neal-zhu 2020-02-21
Sorry, I have not done any tests under the common workload yet. I just ran ``make bench`` 
and of course, there was no big difference compared with the original implementation because normally the worst case won't happen.
I know this optimization will not make a difference in most cases, but I think it won't be a bad idea to replace it with an O(n) algorithm. So I just tried to optimize it while I was reading the source code.
BTW: I'm really excited to receive your reply!