@antirez 2020-02-27
@neal-zhu you are totally right that the O(n) case is not good! The only reason I want to think a bit more about this, is that ziplist.c is going to be totally replaced in the future by listpack.c, so not sure if it is worth to change anything and potentially expose ourselves to some new bug. I'll check it more in depth. Thanks!