@neal-zhu 2020-04-01
Dear sir:
   I add a basic benchmark test for optimized __ziplistCascadeUpdate. The result of the un-optimized version:
```shell
benchmark __ziplistCascadeUpdat elapsed 68
```
while after the optimization result is 
```shell
benchmark __ziplistCascadeUpdat elapsed 0
``` 
  Please take a look at it if you could