@oranagra 2020-08-13
@neal-zhu sorry for not responding to your earlier messages. i intend to review this soon since i'm working on other areas in ziplist at the moment which may de-stabilize it anyway, so i hope to get both shipped into the same redis version.