@oranagra 2020-08-16
@neal-zhu I reviewed the code and i'd like to ask for a few changes:
1. Please add some comments. before each block of a few lines, add a short line that describes what the block does. specifically before any loop or `if` scope. this will make the code much easier to review and edit.
2. Other places in this file clearly make an effort to declare all the variables at the top. although redis does violate this rule in a few places, but this function stands out when compared with others in that file.
3. Correct me if i'm wrong, but the building of that list and all the heap allocations that come with it are not really necessary, the ziplist can be "walked" tail to head, and if we keep track of some delta to skip after each memmove, i think we can do all of that without the additional list. doing so may improve efficiency further since it'll reduce all the heap allocations and improve cache locality.