@neal-zhu 2020-08-17
@oranagra hi, about the third question, do you mean ``entries`` could be eliminated?