@oranagra 2020-08-17
@neal-zhu i'm predicting that the extra heap allocations (2 per items), and reads from them (2 indirections and possible cache misses), make more damage to performance than decoding the header again (considering that memory is being accessed anyway, so it should be in the cache)