@oranagra 2020-08-18
@neal-zhu thank you. i feel this is better. how do you feel?
did you happen to compare the performance to the previous implementation?
can you please run both the redis test suite, and the ziplist test (the one you modified) with valgrind.
for the redis test suite: `make distclean; make valgrind; ./runtest --valgrind`